# docs/swagger.yaml
definitions:
...
  models.AuthRequest:
    properties:
      token:
        type: string
    type: object
...
  models.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 60
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - student
        - teacher
        - admin
        type: string
      surname:
        maxLength: 60
        minLength: 3
        type: string
    required:
    - email
    - name
    - password
    - role
    - surname
    type: object
  models.EmailVerifiaction:
    properties:
      pin:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  utils.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
      title:
        type: string
    type: object
...
host: https://user-api-production-99c2.up.railway.app/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User API
  version: "1.0"
paths:
  /auth/admins:
    post:
      consumes:
      - application/json
      description: Registers a new Admin
      parameters:
      - description: User Registration Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Register a new Admin
      tags:
      - Auth
  /auth/google:
    post:
      consumes:
      - application/json
      description: Authenticate a user using Google OAuth2
      parameters:
      - description: Google OAuth2 Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Authenticate with Google
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with email and password, returning user information
        and a JWT token.
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login returns user and token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or input validation failed
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "403":
          description: User is blocked
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error (e.g., token generation failed)
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: User login
      tags:
      - Auth
...
  /auth/users:
    post:
      consumes:
      - application/json
      description: Registers a new user
      parameters:
      - description: User Registration Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Register a new user
      tags:
      - Auth
  /auth/users/verify:
    post:
      consumes:
      - application/json
      description: Verify the new user's registration using a Pin sent to the users
        email
      parameters:
      - description: Email Verification Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EmailVerifiaction'
      produces:
      - text/plain
      responses:
        "201":
          description: User Verified and created successfully
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Verify a new user's registration
      tags:
      - Auth
  /auth/users/verify/resend:
    put:
      consumes:
      - application/json
      description: Sends a new Verification Pin to email saved in verification cookie
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: New Pin sent successfully
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Sends a new Verification
      tags:
      - Auth
  /auth/verify:
    get:
      description: Verifies the JWT token and returns a success message if valid
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token is valid
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Verify a JWT token
      tags:
      - Auth
...
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"